backend/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-dotenv==1.0.0
# langchain-groq
# langchain
# pydantic==2.5.0
# reportlab==4.0.7
# pandas==2.1.4
# python-multipart==0.0.6
# cors==1.0.1
# fastapi-cors==0.0.1

fastapi
uvicorn
python-dotenv
langchain-groq
langchain
pydantic
reportlab
pandas
python-multipart
cors
fastapi-cors